Move h.auth.worker to h.tasks.auth
